requirements.txt
decord==0.6.0
numpy==2.2.6
pandas==2.3.3
Pillow==11.3.0
torch
tqdm
open_clip_torch
